WIP: Implement LFS plugin storing large objects in an Amazon S3 bucket
This plugin uses JGit's S3Repository class to provide a large file
storage using an Amazon S3 bucket.

This change makes use of the corresponding WIP in JGit [1]. Therefore,
to test locally you need to build JGit from [1] locally. In order to
make this plugin usable you must build Gerrit from [2] with this version
of JGit. This can be done using the multi-cell buck build introduced
in [3]. Follow the instructions given in [3] to build gerrit including
jgit.

The plugin supports only in-tree buck build. To build it clone the
repository under .../plugins folder and then execute:

$ buck build plugins/lfs-storage-s3

[1] https://git.eclipse.org/r/#/c/52177
[2] https://gerrit-review.googlesource.com/72891
[3] https://gerrit-review.googlesource.com/#/c/73000

Change-Id: I635c1c5454a0b23a8e9f9fceab9743692f7c72cb
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>