Update patch set 10
Patch Set 10:

> One suggestion, next time you are doing such an introusive change in a new  patch set ...

Oh, I thought that this will be clear from the changed commit message in the corresponding core change [1]. Basically, now the core change just exposes
the LFS protocol endpoint but doesn't handle it. Instead it forwards all LFS
protocol requests to the plugins configured under lfs.plugin in gerrit.config.

Because of that, all dependencies to JGit's LFS support are removed from the change in core.
It doesn't need them as it only forwards the requests to the configured LFS plugin.

> So I'm clueless why whould expect that the referenced artifacts would be found?
Sorry, I didn't want to make you responsible for that, just asked for any advice.
The LFS artifacts removed from core are now included in this change. This is why I hoped that
the cross-cell support will also work when these artefacts are referenced from here.
So, cross-cell support is not working from in-tree plugin build?

[1] https://gerrit-review.googlesource.com/#/c/72891/11..12//COMMIT_MSG

Patch-set: 10